ace/api/requirements.txt
# Core web framework
Flask==3.0.3
Flask-Cors==4.0.0

# Environment variable management
python-dotenv==1.0.1

# Supabase client
supabase==2.9.0

# OpenAI API
openai==1.55.3 
httpx==0.27.2

# PDF generation
fpdf2==2.7.9

# Utility libraries
requests==2.32.3
typing-extensions>=4.10.0
PyJWT